Decrease memcpy use in win_textbuffer_redraw()
Text selection is implemented by temporarily inverting the "reverse"
attribute on the selected text. To allow temporary modification,
a buffer is used into which the current line is copied.

The original code would copy every line into that buffer, even if the
line contains no selected text and even if the line did not require
repainting at all.

The new code will only allocate and fill the buffer in cases when
selected text is being painted. Also, it will only copy the attribute
indices that are actually in use, instead of the entire line struct.